Add 'Current User Groups Read Current Post' condition
Check if current user can read the current post based on the Groups read access setting. With the use of other Groups filters, this can allow a CPT to be partially or fully visible based on that Groups read access setting. Will document that other aspect in a future update.